requirements.txt
bcrypt==4.0.1
black==22.12.0
click==8.1.3
colorama==0.4.6
fastapi==0.109.0
pydantic==2.5.3
passlib==1.7.4
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
SQLAlchemy==2.0.25
sqlmodel==0.0.14
uvicorn==0.20.0
psycopg-binary==3.1.17
email-validator==2.1.0.post1
psycopg2==2.9.5